+++ packages/core/src/modules/cell.ts
@@ -390,7 +390,7 @@ export function setCellValue(
         const format = getNumberFormat(strValue, commaPresent);
 
         cell.m = v.m ? v.m : update(format, cell.v);
-        cell.ht = v?.ct ? cell.ht : 2;
+        cell.ht = v?.ht ? cell.ht : 2;
         cell.ct = { fa: format, t: "n" };
         if (cell.v === Infinity || cell.v === -Infinity) {
           cell.m = cell.v.toString();