+++ packages/react/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @fileverse-dev/fortune-react
 
+## 1.1.90
+
+### Patch Changes
+
+- ghost position, insert for header
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.1.90
+
+## 1.1.89
+
+### Patch Changes
+
+- data block refresh
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.1.89
+
+## 1.1.88
+
+### Patch Changes
+
+- update formula and cell in between drag rage
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.1.88
+
 ## 1.1.87
 
 ### Patch Changes